requirements.txt
accelerate==0.30.0 
aiohttp==3.9.5 
aiosignal==1.3.1 
altair==5.3.0 
annotated-types==0.6.0 
anyio==4.3.0 
asgiref==3.8.1 
async-timeout==4.0.3 
attrs==23.2.0 
backoff==2.2.1 
bcrypt==4.1.3 
beautifulsoup4==4.12.3 
blinker==1.8.2 
boto3==1.34.100 
botocore==1.34.100 
build==1.2.1 
cachetools==5.3.3 
certifi==2024.2.2 
charset-normalizer==3.3.2 
chroma-hnswlib==0.7.3 
chromadb==0.5.0 
click==8.1.7 
colorama==0.4.6 
coloredlogs==15.0.1 
dataclasses-json==0.6.5 
Deprecated==1.2.14 
dirtyjson==1.0.8 
distro==1.9.0 
dnspython==2.6.1 
email-validator==2.1.1 
exceptiongroup==1.2.1 
faiss-cpu==1.8.0 
fastapi==0.111.0 
fastapi-cli==0.0.2 
filelock==3.14.0 
flatbuffers==24.3.25 
free-proxy==1.1.1 
frozenlist==1.4.1 
fsspec==2024.3.1 
gitdb==4.0.11 
GitPython==3.1.43 
google==3.0.0 
google-ai-generativelanguage==0.4.0 
google-api-core==2.19.0 
google-auth==2.29.0 
google-generativeai==0.4.1 
googleapis-common-protos==1.63.0 
graphviz==0.20.1 
greenlet==3.0.3 
groq==0.5.0 
grpcio==1.63.0 
grpcio-status==1.62.2 
h11==0.14.0 
html2text==2020.1.16 
httpcore==1.0.5 
httptools==0.6.1 
httpx==0.27.0 
huggingface-hub==0.20.3 
humanfriendly==10.0 
idna==3.7 
importlib-metadata==7.0.0 
importlib-resources==6.4.0 
intel-openmp==2021.4.0 
Jinja2==3.1.4 
jmespath==1.0.1 
joblib==1.4.2 
jsonpatch==1.33 
jsonpointer==2.4 
jsonschema==4.22.0 
jsonschema-specifications==2023.12.1 
kubernetes==29.0.0 
langchain==0.1.14 
langchain-aws==0.1.3 
langchain-community==0.0.37 
langchain-core==0.1.52 
langchain-google-genai==1.0.1 
langchain-groq==0.1.3 
langchain-openai==0.1.1 
langchain-text-splitters==0.0.1 
langsmith==0.1.54 
llama-index==0.10.34 
llama-index-agent-openai==0.2.3 
llama-index-cli==0.1.12 
llama-index-core==0.10.34 
llama-index-embeddings-huggingface==0.2.0 
llama-index-embeddings-openai==0.1.9 
llama-index-indices-managed-llama-cloud==0.1.6 
llama-index-legacy==0.9.48 
llama-index-llms-groq==0.1.3 
llama-index-llms-huggingface==0.1.5 
llama-index-llms-openai==0.1.16 
llama-index-llms-openai-like==0.1.3 
llama-index-multi-modal-llms-openai==0.1.5 
llama-index-program-openai==0.1.6 
llama-index-question-gen-openai==0.1.3 
llama-index-readers-file==0.1.20 
llama-index-readers-llama-parse==0.1.4 
llama-index-vector-stores-chroma==0.1.8 
llama-parse==0.4.2 
llamaindex-py-client==0.1.19 
lxml==5.2.1 
markdown-it-py==3.0.0 
MarkupSafe==2.1.5 
marshmallow==3.21.2 
mdurl==0.1.2 
minify-html==0.15.0 
mkl==2021.4.0 
mmh3==4.1.0 
monotonic==1.6 
mpmath==1.3.0 
multidict==6.0.5 
mypy-extensions==1.0.0 
nest-asyncio==1.6.0 
networkx==3.3 
nltk==3.8.1 
numpy==1.26.4 
oauthlib==3.2.2 
onnxruntime==1.17.3 
openai==1.26.0 
opentelemetry-api==1.24.0 
opentelemetry-exporter-otlp-proto-common==1.24.0 
opentelemetry-exporter-otlp-proto-grpc==1.24.0 
opentelemetry-instrumentation==0.45b0 
opentelemetry-instrumentation-asgi==0.45b0 
opentelemetry-instrumentation-fastapi==0.45b0 
opentelemetry-proto==1.24.0 
opentelemetry-sdk==1.24.0 
opentelemetry-semantic-conventions==0.45b0 
opentelemetry-util-http==0.45b0 
orjson==3.10.3 
overrides==7.7.0 
packaging==23.2 
pandas==2.0.3 
pillow==10.3.0 
pip==24.1.2 
playwright==1.43.0 
posthog==3.5.0 
proto-plus==1.23.0 
protobuf==4.25.3 
psutil==5.9.8 
pyarrow==16.0.0 
pyasn1==0.6.0 
pyasn1-modules==0.4.0 
pydantic==2.7.1 
pydantic-core==2.18.2 
pydeck==0.9.0 
pyee==11.1.0 
Pygments==2.18.0 
pypdf==4.2.0 
PyPDF2==3.0.1 
PyPika==0.48.9 
pyproject-hooks==1.1.0 
pyreadline3==3.4.1 
python-dateutil==2.9.0.post0 
python-dotenv==1.0.1 
python-multipart==0.0.9 
pytz==2024.1 
PyYAML==6.0.1 
referencing==0.35.1 
regex==2024.4.28 
requests==2.31.0 
requests-oauthlib==2.0.0 
rich==13.7.1 
rpds-py==0.18.1 
rsa==4.9 
s3transfer==0.10.1 
safetensors==0.4.3 
scikit-learn==1.4.2 
scipy==1.13.0 
scrapegraphai==0.9.0 
sentence-transformers==2.7.0 
setuptools==65.5.0 
shellingham==1.5.4 
six==1.16.0 
smmap==5.0.1 
sniffio==1.3.1 
soupsieve==2.5 
SQLAlchemy==2.0.30 
starlette==0.37.2 
streamlit==1.34.0 
streamlit-chat==0.1.1 
striprtf==0.0.26 
sympy==1.12 
tbb==2021.12.0 
tenacity==8.2.3 
text-generation==0.7.0 
threadpoolctl==3.5.0 
tiktoken==0.5.2 
tokenizers==0.19.1 
toml==0.10.2 
tomli==2.0.1 
toolz==0.12.1 
torch==2.3.0 
tornado==6.4 
tqdm==4.66.3 
transformers==4.40.2 
typer==0.12.3 
typing-extensions==4.11.0 
typing-inspect==0.9.0 
tzdata==2024.1 
ujson==5.9.0 
urllib3==2.2.1 
uvicorn==0.29.0 
watchdog==4.0.0 
watchfiles==0.21.0 
websocket-client==1.8.0 
websockets==12.0 
wrapt==1.16.0 
yarl==1.9.4 
zipp==3.18.1
